fix : ID Long type 변경

+++ src/main/java/knu/chcse/knucseofficialserver/domain/entity/comment/CommentJpaRepository.java
@@ -2,5 +2,5 @@ package knu.chcse.knucseofficialserver.domain.entity.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentJpaRepository extends JpaRepository<Comment, Integer> {
+public interface CommentJpaRepository extends JpaRepository<Comment, Long> {
 }
